refactor: add file VerseUtils foк getLanguageGroups
